+++ packages/plugin/src/out-file.ts
@@ -14,6 +14,9 @@ import {InternalOptions} from "./our-options";
 export class OutFile extends TypescriptFile implements GeneratedFile {
 
 
+    private header: string | undefined;
+
+
     constructor(
         name: string,
         public readonly fileDescriptor: FileDescriptorProto,
@@ -28,16 +31,39 @@ export class OutFile extends TypescriptFile implements GeneratedFile {
         if (this.isEmpty()) {
             return "";
         }
+        return this.getHeader() + super.getContent();
+    }
+
+
+    getHeader(): string {
+        if (this.isEmpty()) {
+            return "";
+        }
+        if (!this.header) {
+            this.header = this.makeHeader();
+        }
+        return this.header;
+    }
+
+
+    private makeHeader(): string {
         let props = [];
         if (this.fileDescriptor.package) {
             props.push('package "' + this.fileDescriptor.package + '"');
         }
         props.push('syntax ' + (this.fileDescriptor.syntax ?? 'proto2'));
-        let header = [
+        const header = []
+        if (this.options.esLintDisable) {
+            header.push(`/* eslint-disable */`);
+        }
+        header.push(...[
             `// @generated ${this.options.pluginCredit}`,
             `// @generated from protobuf file "${this.fileDescriptor.name}" (${props.join(', ')})`,
             `// tslint:disable`
-        ];
+        ]);
+        if (this.options.tsNoCheck) {
+            header.push(`// @ts-nocheck`);
+        }
         if (this.registry.isExplicitlyDeclaredDeprecated(this.fileDescriptor)) {
             header.push('// @deprecated');
         }
@@ -49,7 +75,7 @@ export class OutFile extends TypescriptFile implements GeneratedFile {
         if (head.length > 0 && !head.endsWith('\n')) {
             head += '\n';
         }
-        return head + super.getContent();
+        return head;
     }